+++ packages/framework/src/hooks/__tests__/useRegisterBindings.test.ts
@@ -3,7 +3,7 @@ import { renderHook } from "@testing-library/react";
 import { getDefaultStore } from "jotai";
 import { useRegisterBindings } from "../useRegisterBindings";
 import { screenAtom } from "../../state";
-import { screenStorageAtom } from "../useEnsembleStorage";
+import { storageAtom } from "../useEnsembleStorage";
 
 const mockInvokable = {
   id: "test",
@@ -42,7 +42,7 @@ test("instantiates state from props", () => {
 });
 
 test("evaluates nested bindings", () => {
-  store.set(screenStorageAtom, {
+  store.set(storageAtom, {
     paddingValue: 2,
   });
   store.set(screenAtom, {
