+++ packages/framework/src/hooks/useEnsembleStorage.ts
@@ -1,77 +1,40 @@
-import { atom, useAtom, type WritableAtom } from "jotai";
-import { clone, merge } from "lodash-es";
+import { useAtom } from "jotai";
+import { clone } from "lodash-es";
 import { useMemo } from "react";
-import { type EnsembleStorage } from "../shared/ensemble";
+import { atomWithSessionStorage } from "../state";
 
-// Custom storage atom so it's writable and hydrate-able so values remain consistent
-const atomWithSessionStorage = <T = unknown>(
-  key: string,
-  initialValue: T,
-): WritableAtom<T, unknown[], unknown> => {
-  const getInitialValue = (): T => {
-    const item = sessionStorage.getItem(key);
-    if (item !== null) {
-      return JSON.parse(item) as T;
-    }
-    return initialValue;
-  };
-  const baseAtom = atom<T>(getInitialValue());
-  const derivedAtom = atom<T, unknown[], unknown>(
-    (get) => get(baseAtom),
-    (get, set, update) => {
-      const nextValue = (
-        typeof update === "function" ? update(get(baseAtom)) : update
-      ) as T;
-      set(baseAtom, nextValue);
-      sessionStorage.setItem(key, JSON.stringify(nextValue));
-    },
-  );
-  return derivedAtom;
-};
-
-export const screenStorageAtom = atomWithSessionStorage<{
+export interface EnsembleStorage {
   [key: string]: unknown;
-}>("ensemble.storage", {});
+}
 
-export const useEnsembleStorage = (): EnsembleStorage => {
-  const [storage, setStorage] = useAtom(screenStorageAtom);
-  const storageBuffer = useMemo<{ [key: string]: unknown }>(() => ({}), []);
+export const storageAtom = atomWithSessionStorage<EnsembleStorage>(
+  "ensemble.storage",
+  {},
+);
 
-  useMemo(() => {
-    merge(storageBuffer, storage);
-  }, [storageBuffer, storage]);
+interface EnsembleStorageBuffer {
+  set: (key: string, value: unknown) => void;
+}
 
-  const storageApi = useMemo(
-    () => createStorageApi(storageBuffer, setStorage),
-    [setStorage, storageBuffer],
+export const useEnsembleStorage = (): EnsembleStorage &
+  EnsembleStorageBuffer => {
+  const [storage, setStorage] = useAtom(storageAtom);
+
+  const storageBuffer = useMemo<EnsembleStorageBuffer>(
+    () => ({
+      set: (key: string, value: unknown): void => {
+        storage[key] = value;
+        setStorage(clone(storage));
+        window.dispatchEvent(
+          new StorageEvent("storage", { key: "ensemble.storage" }),
+        );
+      },
+      get: (key: string): unknown => {
+        return storage[key];
+      },
+    }),
+    [setStorage, storage],
   );
 
-  return storageApi;
-};
-
-export const createStorageApi = (
-  storage?: { [key: string]: unknown },
-  setStorage?: (storage: { [key: string]: unknown }) => void,
-): EnsembleStorage => {
-  return {
-    set: (key: string, value: unknown): void => {
-      if (storage) {
-        storage[key] = value;
-        setStorage?.(clone(storage));
-      }
-    },
-    get: (key: string): unknown => {
-      return storage?.[key];
-    },
-    delete: (key: string): unknown => {
-      const oldVal = storage?.[key];
-      if (storage) {
-        // eslint-disable-next-line @typescript-eslint/no-dynamic-delete
-        delete storage[key];
-        setStorage?.(clone(storage));
-      }
-      return oldVal;
-    },
-    ...storage,
-  };
+  return { ...storageBuffer, ...storage };
 };
